requirements.txt
Django>=4.2,<5.0
markdown>=3.5.1
python-decouple>=3.8
Pillow>=10.0.0
gunicorn>=21.2.0